ref(cardinality): Split redis module into multiple smaller modules
